backend/timeframe_parser/requirements.txt
fastapi
uvicorn
python-dateutil
supabase
python-dotenv
